Memory Saved Trick #1
Hold times in unsigned int instead of long and handle the rollover.
Saves 4 byte per trigger = 80 bytes.

+++ Arduino/v3/SensactArduinoCode/Triggers.h
@@ -25,13 +25,13 @@ class Trigger {
     char actionID;
     char actionState;
     long actionParameters;
-    int  delayMs;
+    unsigned int  delayMs;
     char repeat;
     
     // Dynamic elements
     boolean actionTaken;
-    long onTime;
-    long lastActionTime;
+    unsigned int onTime;
+    unsigned int lastActionTime;
     int repeatInterval;
     int repeatCount;
